real_estate_analyzer/requirements.txt
requests>=2.31.0
pandas>=2.0.0
dash>=2.14.0
plotly>=5.15.0
numpy>=1.24.0
scipy>=1.11.0 
python-dotenv>=1.0.0 